Feat: 게시판 조회, 상세 조회, 삭제 기능 등

+++ backend/src/main/java/com/ssafy/enjoytrip/features/tripost/application/dto/TripostListItemDto.java
@@ -20,9 +20,9 @@ public class TripostListItemDto {
     private String uid;
     private String name;
     private LocalDateTime createdAt;
-    private Long commentCount;
-    private Long viewCount;
-    private Long likeCount;
+    private Integer commentCount;
+    private Integer viewCount;
+    private Integer likeCount;
     private List<WaypointSnapshotDto> waypoints;
     private List<WaypointImageDto> images;
 }
